Tests can't be run in productions.
Ember doesn't include the test files for production builds.
Additionally, the order in which files are being included requires
that Ember be included, then some test helpers, then the application
code. On the client side, this restricts things. On the server side,
it's just a copy of files, so why re-test? The only difference would
be setting an environment variable, but for that the server should
run in TEST mode anyway.

For now, production tests won't work.